+++ Source/Backend/OnlineStore/OnlineStore.API.Identity/Domain/AuthDbContext.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace OnlineStore.API.Identity.Models
+namespace OnlineStore.API.Identity.Domain
 {
     public class AuthDbContext : DbContext
     {
