+++ Source/Backend/OnlineStore/OnlineStore.API.Identity/Domain/AuthDbContextExtentions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using IdentityModel;
 
-namespace OnlineStore.API.Identity.Models
+namespace OnlineStore.API.Identity.Domain
 {
     public static class AuthDbContextExtentions
     {
