+++ Source/Backend/OnlineStore/OnlineStore.API.Identity/Domain/UserClaim.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace OnlineStore.API.Identity.Models
+namespace OnlineStore.API.Identity.Domain
 {
     public class UserClaim
     {
