+++ Source/Backend/OnlineStore/OnlineStore.API.Identity/Services/ProfileService.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using IdentityServer4.Extensions;
 using IdentityServer4.Models;
 using IdentityServer4.Services;
-using OnlineStore.API.Identity.Models;
+using OnlineStore.API.Identity.Domain;
 
 namespace OnlineStore.API.Identity.Services
 {
